+++ jenna/material/core/textureLoader.js
